src/torch_process_logistic/requirements.txt
pandas
nltk
joblib
regex
scikit-learn
tqdm
mlflow
python-dotenv
torch